+++ packages/legacy/utils/src/selectionSet.ts
