+++ packages/legacy/utils/test/selectionSet.spec.ts
@@ -1,5 +1,5 @@
 import { parseSelectionSet } from '@graphql-tools/utils';
-import { containsSelectionSet } from '../src/containsSelectionSet';
+import { containsSelectionSet } from '../src/selectionSet';
 
 it.each([
   {
